### 2024-02-26 #235 @jatkinson1000
Hi @labarba Thanks for the update and I understand.

We have delivered this workshop a few times now so there have been incremental updates to the repository, but the paper should still be fine as-is (subject to review).
I'll tag my co-author @jdenholm in case he has anything to add.
