### 2024-03-01 #235 @nicoguaro
@manubastidas, @mnarayan, would you be interested in reviewing this submission for JOSE?
It concerns a workshop that "Provide participants with the ability to develop ML models in PyTorch" as the main learning outcome.
