### 2024-03-01 #235 @mnarayan
Sure, I can do it.

On Thu, Feb 29, 2024, 10:38 PM Nicolás Guarín-Zapata <
***@***.***> wrote:

> @manubastidas <https://github.com/manubastidas>, @mnarayan
> <https://github.com/mnarayan>, would you be interested in reviewing this
> submission for JOSE?
> It concerns a workshop with "Provide participants with the ability to
> develop ML models in PyTorch" as the main learning outcome.
>
