### 2024-03-03 #235 @manubastidas
I agree with reviewing it

Best regards,

On Thu, 29 Feb 2024 at 22:38, Nicolás Guarín-Zapata <
***@***.***> wrote:

> @manubastidas <https://github.com/manubastidas>, @mnarayan
> <https://github.com/mnarayan>, would you be interested in reviewing this
> submission for JOSE?
> It concerns a workshop with "Provide participants with the ability to
> develop ML models in PyTorch" as the main learning outcome.
>
