### 2024-03-03 #239 @editorialbot
[REVIEW]: Practical machine learning with PyTorch
**Submitting author:** @jatkinson1000 (<a href="http://orcid.org/0000-0001-5001-4812">Jack Atkinson</a>)
**Repository:** https://github.com/Cambridge-ICCS/ml-training-material
**Branch with paper.md** (empty if default branch): JOSE
**Version:** v1.0
