### 2024-03-03 #239 @nicoguaro
@manubastidas, @mnarayan,  this is the space where the review process takes form. There is a checklist for each one, tick the boxes when you see that the criterion is satisfied. You can generate the checklist with

```@editorialbot generate my checklist```

I will be here to answer the questions that you might have.

Let us use as a tentative timeframe the first week of April, is that OK for you?
