### 2024-03-26 e7dccf5
:memo: Type-hint and docstring fix (#44)
