### 2024-04-18 #239 @nicoguaro
I think that @manubastidas is not able to continue the review with us for personal reasons. Thank you for your disposition and I hope to count with you in future opportunities.

@dortiz5 has accepted to help us with the review.
