### 2024-04-18 #239 @nicoguaro
@dortiz5, this is the space where the review process takes form. There is a checklist for each one, tick the boxes when you see that the criterion is satisfied. You can generate the checklist with

```@editorialbot generate my checklist```

I will be here to answer the questions that you might have.

Let us use as a tentative timeframe the second week of May, is that OK for you?
