### 2024-05-06 570df1b
Add contribution and support guidelines to the README.
