### 2024-05-06 #239 @jatkinson1000
Thanks for looking at this @dortiz5 !

I have just opened a Pull Request to add contribution guidelines [here](https://github.com/Cambridge-ICCS/ml-training-material/pull/48).

Once that has been reviewed I'll merge it to main, and then rebase the paper branch off of main and let you know.
