### 2024-05-06 #239 @dortiz5
@jatkinson1000, thanks to you for the repository.

It is nicely written, ordered, and storytelling, with clear instructions for preparation and prerequisites to run the exercises. Also, the local installation was easy. I followed the four simple steps you described, and it works correctly.

From the pedagogical standpoint, learning objectives are explicitly written and aligned with the content in the slides and the exercises. I liked that you gave different options to run the exercises and gave their solutions. Also, the information is concise and easy to follow. Finally, I found that the JOSE paper follows the checklist.
