### 2024-05-09 cab7ff6
Revert "Add binder info to README."
This reverts commit 1d6a70843f1f8137f32f5954b920a6924cca58d8.
