### 2024-05-09 4789f11
Revert "Add setup.py file to allow running on binder."
This reverts commit c1e33f6f2e69a13d649d196d67aebdc8b796203b.
