+++ setup.py
@@ -1,6 +1,0 @@
-#!usr/bin/env python
-
-from setuptools import setup
-
-if __name__ == "__main__":
-    setup()
