### 2024-05-09 7fc7646
Revert ":memo: Type-hint and docstring fix (#44)"
This reverts commit 0fc508967a6744193f52cc5a93ea1aa98f2d5ef7.
