+++ src/ml_workshop/_penguins.py
@@ -1,5 +1,4 @@
 """Penguins dataset."""
-
 from typing import Optional, List, Dict, Tuple, Any
 
 from torch.utils.data import Dataset
@@ -18,9 +17,9 @@ class PenguinDataset(Dataset):
 
     Parameters
     ----------
-    input_keys : List[str]
+    input_keys : Sequence[str]
         The column titles to use in the input feature vectors.
-    target_keys : List[str]
+    target_keys : Sequnce[str]
         The column titles to use in the target feature vectors.
     train : bool
         If ``True``, this object will serve as the training set, and if
@@ -40,7 +39,7 @@ class PenguinDataset(Dataset):
     def __init__(
         self,
         input_keys: List[str],
-        target_keys: List[str],
+        target_keys: str,
         train: bool,
         x_tfms: Optional[Compose] = None,
         y_tfms: Optional[Compose] = None,
