### 2024-05-09 930d4f6
Revert "Update githubb pages site with binder and small readme phrasing."
This reverts commit 9b3f39df44ebf216e0c9fd1ab293790806030212.
