### 2024-05-09 3ca2f9a
Revert "Small updates for NCAS summer school improving background info (#24)"
This reverts commit fc808a8b986fb0ef87ae965386e3e3a372a8d46b.
