### 2024-05-09 2c8b0a1
Revert "Add contribution and support guidelines to the README. (#48)"
This reverts commit 8a220625738f12af8a6602442230d87608ed4852.
