### 2024-05-09 465a52b
Revert "Update README moving colab solutions and adding binder info."
This reverts commit 8813a5f19c85f1fbcb3673c068614d9f30b41518.
