### 2024-05-09 #239 @jatkinson1000
Hi @dortiz5 There should now be contribution guidelines on the main branch and the paper branch.

I'll look at sorting a version and let you know once that's done.

@mnarayan @nicoguaro please do let me know if there is anything I can do to help you.
Thanks all!
