### 2024-05-17 #239 @jatkinson1000
@dortiz5 Thank you so much for your time and kind words.

@nicoguaro and @mnarayan please let us know if there is anything at all I can do to get this over the finish line and accepted, especially after @dortiz5's positive review. We are delivering this course again in July and it would be great to be able to feature/link to JOSE at the workshop.
