### 2024-05-29 Cambridge-ICCS/ml-training-material#53 @pdpino
Broken URL jackatkinson.net
Going through the README I noticed problems with these links:

* the URL [jackatkinson.net](https://jackatkinson.net/) is not working. If the website does not exist anymore, I'd suggest removing the link from the README

![image](https://github.com/Cambridge-ICCS/ml-training-material/assets/20805451/0a462f51-9db2-46d0-b6a8-0c95ecad2bc7)

* There is a mention to "existing open issue" in the [Contribution Guidelines](https://github.com/Cambridge-ICCS/ml-training-material#contribution-guidelines-and-support) with a link to [open a new issue](https://github.com/Cambridge-ICCS/ml-training-material/issues/new/choose), but I'd say it should be a link to [issues](https://github.com/Cambridge-ICCS/ml-training-material/issues)
