### 2024-05-29 Cambridge-ICCS/ml-training-material#56 @pdpino
Suggestion: include used library versions
I'd suggest mentioning which versions you used for major python libraries, especifically pytorch, matplotlib, pandas, numpy. For example, just saying: "The exercises were tested with versions: pytorch=2.2.2, pandas=2.1.0, etc".

In general, I think this is a good idea for future-proofing the repository -- is useful for users to know which versions should work. (For me it's ok that versions are not specified in the pyproject.toml, in some cases I've found problems when too many packages and versions are specified)
