### 2024-05-29 #239 @pdpino
Hello everyone, sorry for the late reply, I finally found some time to check this out.

Overall it looks really good! I think the slides and exercises are very pedagogic. I left some minor suggestions and changes (https://github.com/Cambridge-ICCS/ml-training-material/issues/53, https://github.com/Cambridge-ICCS/ml-training-material/pull/54, https://github.com/Cambridge-ICCS/ml-training-material/pull/55, https://github.com/Cambridge-ICCS/ml-training-material/issues/56), but these are non-blocking and I recommend the work for publication.

@nicoguaro what does the point about Version mean in this case? There are no releases in this repository, so I'd say it doesn't apply in this case. @jatkinson1000 though I might suggest you version your releases, in case you want to change this course in the future.
