### 2024-05-30 7eccfe3
Add major package versions on readme in case users experience issues in future.
