### 2024-05-30 Cambridge-ICCS/ml-training-material#53 @jatkinson1000
Thanks. I think you picked an unfortunate time when the server was down for maintenance, hopefully the link is working now?

The 'issues' issue should be dealt with in #57 if you fancy reviewing that?
