### 2024-05-30 Cambridge-ICCS/practical-ml-with-pytorch#56 @jatkinson1000
This is also now done in #57 if you want to review it there.
