### 2024-05-30 #239 @jatkinson1000
Thanks @pdpino for your kind words.

I have accepted your contributions, and opened a couple of PRs for the issues you raised.

Once these are merged I will apply a version number and release tag ready for publication!
