### 2024-05-30 Cambridge-ICCS/practical-ml-with-pytorch#53 @pdpino
The link is still not working for me. But if is just a temporary issue then no worries!
