### 2024-05-30 Cambridge-ICCS/practical-ml-with-pytorch#53 @jatkinson1000
Hmm, seems I needed to reboot apache.
It should be there now(?), but I'll keep an eye to see if this occurs again in future.
