### 2024-05-30 ad7a431
Update acknowledgments from S. Futures to Sciences.
