### 2024-05-31 5d0fe17
revert black to being an optional linting dependency.
