### 2024-05-31 736e054
Added activation functions to MLP models (and minor formatting). (#45)
* :memo: Added activation functions to model

* :art: Applied black

* :memo: Demo updates

* :art: Applied black

* :heavy_plus_sign: Added black

* :art: Applied black

* revert black to being an optional linting dependency.

* revert explicit specification of learning rate in exercise 1 (discussed in exercise 2.)

* Apply black unrelated to the rest of this PR -_-.

---------

Co-authored-by: Jack Atkinson <jwa34@cam.ac.uk>
