### 2024-05-31 f7d9e1f
Add JOSE Paper (#33)
* Add jose workflow file.

* Add first draft of jose paper

* Address minor comments from @jdenholm.

* Update paper with general comments from @jdenholm - loss fn and encoding.

* Remove summary.

* Add SF link.

* Update quote marks.

* Tidy statement of need.

* Update acknowledgments from S. Futures to Sciences.
