+++ JOSE_paper/paper.md
@@ -1,0 +1,337 @@
+---
+title: 'Practical machine learning with PyTorch'
+tags:
+  - Python
+  - PyTorch
+  - machine learning
+  - Jupyter notebook
+authors:
+  - name: Jack Atkinson
+    orcid: 0000-0001-5001-4812
+    affiliation: "1"
+    corresponding: true
+  - name: Jim Denholm
+    affiliation: "1"
+    orcid: 0000-0002-2389-3134
+affiliations:
+ - name: Institute of Computing for Climate Science, University of Cambridge, UK
+   index: 1
+ # - name: AstraZeneca, UK
+ #   index: 2
+date: 22 October 2023
+bibliography: paper.bib
+
+---
+
+
+# Summary
+
+In the last decade machine learning (ML) and
+deep learning (DL)^[We will use the term ML when talking about both ML and DL in this article]
+have revolutionised many fields within science, industry, and beyond.
+Researchers across domains from the physical sciences to the digital humanities
+are increasingly looking to leverage these tools in their research.
+Many will be experts within their own domains, but will not have
+received any training in machine learning.
+
+We have developed, and delivered, a set of materials entitled _Practical
+machine learning with PyTorch_, designed to teach participants how to _actually_
+write and run ML code in a _hands-on_ fashion whilst also illustrating important design considerations.
+
+
+# Statement of need
+
+With the explosion of ML and DL there have been several promising opportunities
+to apply these techniques in research.
+There are notable applications across many fields from
+the physical sciences [@carleo2019machine], climate science [@kashinath2021physics],
+to the digital humanities [@gefen2021ai].
+
+Whilst there exist many examples of ML code online, it is often in the form of
+complete codes to be downloaded, read, and run by the user.
+These are often missing any discussion of theory, the development process, or
+alternative approaches beyond the scope of the specific example.
+In contrast, much theoretical ML material addresses high-level concepts
+without discussing coding considerations or details of how to
+actually use popular frameworks to implement the models.
+
+Many know how ML works in an abstract sense, but will be unfamiliar with lower-level
+practicalities such as image transforms and other preprocessing techniques required
+to present data to neural networks.
+They can describe how something works, but would have no idea where to start if
+asked to do it.
+Such practical aspects are ideally learnt through trial-and-error and
+hands-on experience.
+
+Many machine learning frameworks are accessed using a Python framework.
+One such commonly used framework is [PyTorch](https://pytorch.org/) [@paszke2019pytorch].
+Researchers are likely to have experience writing Python code, but not PyTorch.
+
+
+# Learning objectives
+
+The key learning objective from this workshop could be simply summarised as:  
+_"Provide participants with the ability to develop ML models in PyTorch"_.
+
+However, there are a few subtleties that we wish to highlight.
+We go beyond the ability to blindly run downloaded code to:
+
+* provide an understanding of the structure of a PyTorch model and ML pipeline,
+* introduce the different functionalities PyTorch might provide,
+* encourage good research software engineering (RSE) practice, and
+* exercise careful consideration and understanding of data used for training ML models.
+
+With regards to specific ML content we cover:
+
+* using ML for both classification and regression,
+* artificial neural networks (ANNs) and convolutional neural networks (CNNs), and
+* treatment of both tabular and image data.
+
+
+# Teaching materials
+
+All of the teaching materials for this course are available online in a GitHub repository.
+In addition we have a
+[GitHub pages site](https://cambridge-iccs.github.io/ml-training-material/)
+as a central resource to point participants to.
+
+## Slides
+
+We have produced two slide decks for the course, both available online and linked from
+both the repository and the GitHub pages site.
[…]
