### 2024-05-31 2b7bdb9
Version 1.0.0 and add metadata to pyproject.toml.
