### 2024-05-31 #239 @jatkinson1000
@nicoguaro

I have merged all of the pull requests and created a tagged release v1.0.0 on Github.
This can be viewed here: https://github.com/Cambridge-ICCS/practical-ml-with-pytorch

Thank you (and the reviewers) for taking us through the process.
