### 2024-05-31 #239 @jatkinson1000
@nicoguaro I have also created a Zenodo release for v1.0.0 here: https://zenodo.org/records/11401113
