### 2024-06-04 #239 @nicoguaro
@labarba, this paper is ready to move forward.
