### 2024-06-04 #239 @jatkinson1000
@nicoguaro This is amazing, thank you for this.

I have just checked the final proof and some of the links appear to be old (and no longer working).
This is perhaps because we have updated them after our original submission.
The latest version of the paper that was reviewed can be found on our main branch here: https://github.com/Cambridge-ICCS/practical-ml-with-pytorch/tree/main/JOSE_paper
I have also re-made the `JOSE` branch (merged this week) was originally specified as the paper branch for the review.

Is it possible to update the final proof to target that?
Apologies for any confusion and issues this may cause.
