### 2024-06-21 #239 @labarba
@jatkinson1000  👋
I found a small typo on page 5 (line 191): "good questions and discussion good…"
Since you are making that fix, might as well also fix capitalization of "palmer" on line 83, and "jupyter" on lines 64 and 67. Also, caps-protect with `{}` the word Jupyter in the title of the reference on line 202. Please also add the DOI to this reference, hosted in preprint form at https://doi.org/10.6084/m9.figshare.19608801
