### 2024-06-21 #239 @labarba
We request that authors edit the metadata of the Zenodo deposit so title and author list match the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object." Could you do that?

Also, I noticed that the ORCID of co-author Jim Denholm is missing from the Zenodo archive. Could you add?
