+++ JOSE_paper/paper.bib
@@ -6,11 +6,13 @@
   year={2019}
 }
 
-@book{barba2019teaching,
-  title={Teaching and learning with Jupyter},
-  author={Barba, Lorena A and Barker, Lecia J and Blank, Douglas S and Brown, Jed and Downey, Allen B and George, Timothy and Heagy, Lindsey J and Mandli, Kyle T and Moore, Jason K and Lippert, David and others},
-  year={2019},
-  url={https://jupyter4edu.github.io/jupyter-edu-book/}
+@article{barba2022teaching,
+  author = "Lorena A. Barba and Lecia J. Barker and Douglas S. Blank and Jed Brown and Allen Downey and Timothy George and Lindsey J. Heagy and Kyle Mandli and Jason K. Moore and David Lippert and Kyle Niemeyer and Ryan Watkins and Richard West and Elizabeth Wickes and Carol Willling and Michael Zingale",
+  title = "{Teaching and Learning with Jupyter}",
+  year = "2022",
+  month = "4",
+  url = "https://figshare.com/articles/online_resource/Teaching_and_Learning_with_Jupyter/19608801",
+  doi = "10.6084/m9.figshare.19608801.v1"
 }
 
 @article{carleo2019machine,
