### 2024-06-22 #239 @jatkinson1000
Hi @labarba I believe I have made those requested changes on the JOSE branch of our repo: https://github.com/Cambridge-ICCS/practical-ml-with-pytorch/commit/48596f5ff7158c5c3eb300554fa3c3b1aeb508e8

I have also updated the zenodo metadata [zenodo.org/11401113](https://zenodo.org/records/11401113), but note that since we include the paper in the archive these new changes will not appear there.

We used the 2019 version of the book when preparing the materials, so I will trust that the content has not significantly changed in the new version.

Please let us know if there is anything else. We are excited to finally see this in print and share.
I hope to make more submissions in future now that the journal is active again.
Thank you.
