### 2024-06-22 #239 @labarba
Thanks, @jatkinson1000.
The deposit on Figshare of the Jupyter4edu book is simply an archive of the PDF export of the same version that was published as an online book originally. (So in fact you could leave the URL field as before and only add the DOi.)
I don't see in the paper diff that you capitalized "Jupyter"?
