### 2024-06-22 be550b5
Capitalise J in Jupyter where appropriate.
