+++ JOSE_paper/paper.bib
@@ -8,7 +8,7 @@
 
 @article{barba2022teaching,
   author = "Lorena A. Barba and Lecia J. Barker and Douglas S. Blank and Jed Brown and Allen Downey and Timothy George and Lindsey J. Heagy and Kyle Mandli and Jason K. Moore and David Lippert and Kyle Niemeyer and Ryan Watkins and Richard West and Elizabeth Wickes and Carol Willling and Michael Zingale",
-  title = "{Teaching and Learning with Jupyter}",
+  title = "{Teaching and Learning with {Jupyter}}",
   year = "2022",
   month = "4",
   url = "https://figshare.com/articles/online_resource/Teaching_and_Learning_with_Jupyter/19608801",
