### 2024-06-22 #239 @jatkinson1000
Thanks @labarba
Ah, I misunderstood - I thought you were referring to capitalisation in the reference.

I have just pushed another commit that capitalises J in Jupyter where appropriate throughout the paper: ~https://github.com/Cambridge-ICCS/practical-ml-with-pytorch/commit/a89e20c867cbcf933f191dc79dba0f331f22059e~ https://github.com/Cambridge-ICCS/practical-ml-with-pytorch/commit/be550b546b925f37446e99bac2d922bbf842a2c2
