### 2024-06-22 #239 @labarba
I see that the teaching materials are double-licensed with MIT for code and CC-BY-NC-SA for content. However, the Zenodo archive indicates MIT and CC-BY. One of these needs to be changed to match...
