### 2024-06-22 #239 @labarba
Hm...from the [landing page](https://cambridge-iccs.github.io/practical-ml-with-pytorch/), under the heading **Google Colab**, when clicking on the link to the first exercise I got an error...
