### 2024-06-22 4488ba4
Fix broken links for colab solutions.
