### 2024-06-22 #239 @jatkinson1000
@labarba Zenodo should now be updated to match the repository.

Colab links should be fixed if you can confirm.
We must have updated the notebook names to append `_solutions` but forgotten to update that particular link.
Thank you for the eagle-eyed spot.

Do we need to re-issue the zenodo release following these changes?
