### 2024-06-23 #239 @labarba
> Do we need to re-issue the zenodo release following these changes?

Up to you. I consider these minor changes, and I assume that over time there could continue to be minor edits on the content, after which at some point you will issue a new archive version. I'm fine to publish now.
